+++ src/Kunc.RiotGames.Lol.DataDragon/Item/StatsDto.cs
@@ -3,13 +3,13 @@
 public class StatsDto : BaseDto
 {
     public double FlatMovementSpeedMod { get; set; }
-    public int FlatHPPoolMod { get; set; }
+    public double FlatHPPoolMod { get; set; }
     public double FlatCritChanceMod { get; set; }
-    public int FlatMagicDamageMod { get; set; }
-    public int FlatMPPoolMod { get; set; }
-    public int FlatArmorMod { get; set; }
-    public int FlatSpellBlockMod { get; set; }
-    public int FlatPhysicalDamageMod { get; set; }
+    public double FlatMagicDamageMod { get; set; }
+    public double FlatMPPoolMod { get; set; }
+    public double FlatArmorMod { get; set; }
+    public double FlatSpellBlockMod { get; set; }
+    public double FlatPhysicalDamageMod { get; set; }
     public double PercentAttackSpeedMod { get; set; }
     public double PercentLifeStealMod { get; set; }
     public double FlatHPRegenMod { get; set; }
